requirements.txt
python-telegram-bot==20.7
gspread==6.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
schedule==1.2.0
pytz==2024.1
python-dotenv==1.0.0
flask==3.0.0
